requirements.txt
flask==3.1.2
pillow
streamlit
tensorflow
numpy